Update to TensorFlow Keras
Update the code to work with TensorFlow Keras. Some of the import
changes
are also likely a result of changing Keras levels.

The original code also assumed a Theano backend to Keras so at least one
of the changes is for Theano to TensorFlow change.

With the move the tensorflow.keras rather than regular 'keras' the
model's use
of an Instance normalization from keras.contrib also had to change.  The
InstanceNormalization class from keras.contrib has been pulled into this
repo
and has retained its original license and copyright.